added logo & particle animation
